requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2>=2.8.6,<2.9
# psycopg2-binary>=2.8.6,<2.9
# Pillow==8.3.1